+++ package.json
@@ -2,11 +2,11 @@
   "name": "noderssbot",
   "version": "0.4.1",
   "description": "Another Telegram RSSBot in Node.js",
-  "main": "index.js",
+  "main": "source/index.js",
   "scripts": {
-    "start": "node -e 'process.env.NODE_PRODUTION = process.env.NODE_PRODUTION || \"production\"; require(\"./index.js\");'",
-    "start-withsnapshot": "node --heapsnapshot-signal=SIGUSR2 -e 'process.env.NODE_PRODUTION = process.env.NODE_PRODUTION || \"production\"; require(\"./index.js\");'",
-    "dev": "proxychains -q node index",
+    "start": "node -e 'process.env.NODE_PRODUTION = process.env.NODE_PRODUTION || \"production\"; require(\"./source/index.js\");'",
+    "start-withsnapshot": "node --heapsnapshot-signal=SIGUSR2 -e 'process.env.NODE_PRODUTION = process.env.NODE_PRODUTION || \"production\"; require(\"./source/index.js\");'",
+    "dev": "proxychains -q node source/index",
     "test": "RSSBOT_DB_PATH=./test/test-data/test.db npx jest --verbose",
     "format": "eslint \"**/*.js\" --fix && prettier \"**/*.{js,json,md}\" --write",
     "lint": "eslint \"**/*.js\" && prettier-check \"**/*.{js,json,md}\" && npm run lint-lockfile",
