+++ source/i18n/ka.yaml
