+++ test/i18n.test.js
@@ -1,4 +1,4 @@
-const i18n = require('../i18n');
+const i18n = require('../source/i18n');
 
 test('i18n', () => {
     expect(i18n).toHaveProperty('zh-cn');
